De-duplicate the PathMap helpers into a shared PathMapParser
The build task carried private copies of the compiler's path-map parse and
prefix-rewrite helpers because Microsoft.Build.Tasks.CodeAnalysis deliberately
does not reference Microsoft.CodeAnalysis. That left two copies of load-bearing,
determinism-critical logic in the repo that had to be kept in sync by hand.

Extract a dependency-free Roslyn.Utilities.PathMapParser and share it as source.
NormalizePathPrefix is generic over the list type so the compiler's ImmutableArray
and the task's List are both consumed without boxing and the task needs no
ImmutableArray reference; ParsePathMap is the diagnostic-free variant the task
needs. PathUtilities.NormalizePathPrefix delegates to it and the task drops its
copies. Shared via the existing source-link pattern used for PathUtilities.cs
itself; BuildHost is excluded by the MSBUILDWORKSPACE_BUILDHOST guard.

The compiler's diagnostic-reporting CommandLineParser.ParsePathMap and the
general EnsureTrailingSeparator / SplitWithDoubledSeparatorEscaping helpers are
left in place: sharing them would fight the BuildHost's minimal surface and lose
the compiler's pooling for no real gain.

Copilot-Session: ce65e9a8-2897-42c7-bb6a-55d59c40bac4
