+++ .github/instructions/Compiler.instructions.md
@@ -81,6 +81,7 @@ Compiler tasks are in `src/Compilers/Core/MSBuildTask/`:
 - `Csc.cs` - C# compiler task
 - `Vbc.cs` - VB compiler task  
 - `ManagedCompiler.cs` - Base compiler task functionality
+- `GenerateMSBuildEditorConfig.cs` - Generates the global analyzer config from `CompilerVisibleProperty`/`CompilerVisibleItemMetadata`. Two **opt-in, default-off** MSBuild properties in `Microsoft.Managed.Core.targets` drive path mapping via the task's `MapSectionHeaderPaths` / `MapPropertyValues` parameters (fed the compiler `$(PathMap)`): `MapGeneratedMSBuildEditorConfigPaths` maps the section-header file paths; `MapGeneratedMSBuildEditorConfigPropertyValues` maps path-valued property values (e.g. `build_property.ProjectDir`, prefix-anchored so non-path/outside-root values are untouched). Both off by default so unopted projects are byte-for-byte unchanged. The task shares the compiler's path-map parse/rewrite logic as source via `src/Compilers/Core/Portable/FileSystem/PathMapParser.cs` (a dependency-free `Roslyn.Utilities.PathMapParser`, `#if !MSBUILDWORKSPACE_BUILDHOST`), which `PathUtilities.NormalizePathPrefix` also delegates to — so there is one implementation, not per-copy "keep in sync" comments. `AnalyzerConfigSet.GetOptionsForSourcePath` applies the map when matching a source path against **global** config sections only (additive: the real path is always tried too, so this is unconditional and safe) — public `AnalyzerConfigSet.Create(pathMap)` overload; the command-line compiler passes `Arguments.PathMap`, the workspace passes `CompilationOptions.SourceReferenceResolver`'s `PathMap` threaded through `ProjectState`. Directory-scoped editorconfig matching always uses the real path.
 
 ## Performance Considerations
