+++ src/Interactive/Host/Microsoft.CodeAnalysis.InteractiveHost.csproj
@@ -51,6 +51,7 @@
     <Compile Include="..\..\Compilers\Core\Portable\InternalUtilities\Hash.cs" Link="Utilities\Hash.cs" />
     <Compile Include="..\..\Compilers\Core\Portable\FileSystem\FileUtilities.cs" Link="Utilities\FileUtilities.cs" />
     <Compile Include="..\..\Compilers\Core\Portable\FileSystem\PathUtilities.cs" Link="Utilities\PathUtilities.cs" />
+    <Compile Include="..\..\Compilers\Core\Portable\FileSystem\PathMapParser.cs" Link="Utilities\PathMapParser.cs" />
     <Compile Include="..\..\Compilers\Core\Portable\FileSystem\PathKind.cs" Link="Utilities\PathKind.cs" />
     <Compile Include="..\..\Compilers\Shared\NamedPipeUtil.cs" Link="Utilities\NamedPipeUtil.cs" />
   </ItemGroup>
